Migrate resizable from /docs/components/ to /components/
Move all examples (Horizontal, Vertical, With Handle, Three Panels) from
the old documentation page to the new reference page, delete the old page,
and update all routes, navigation links, and E2E tests.
