+++ site/ui/renderer.tsx
@@ -104,7 +104,7 @@ const menuEntries: SidebarEntry[] = [
       { title: 'Popover', href: '/docs/components/popover' },
       { title: 'Progress', href: '/docs/components/progress' },
       { title: 'Radio Group', href: '/components/radio-group' },
-      { title: 'Resizable', href: '/docs/components/resizable' },
+      { title: 'Resizable', href: '/components/resizable' },
       { title: 'Scroll Area', href: '/components/scroll-area' },
       { title: 'Select', href: '/components/select' },
       { title: 'Sidebar', href: '/docs/components/sidebar' },
